Refactor settings modals and styles: update import paths and remove unused CSS
