+++ frontend/src/pages/SettingsPages/Modal.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode, useEffect } from "react";
-import "./Modal.css";
+import "./css/Modal.css";
 
 interface ModalProps {
   isOpen: boolean;
